Add seperate rosters page that get's fed from data directory files
Jekyll allows for data files to hold data that you then can use to loop
over during the build. Details here: https://jekyllrb.com/docs/datafiles/

This PR adds a roster page and uses 3 teams data files written in yml to
populate it. This way only the yml files will need to be changed to
update the rosters in the future.